watchlist-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
motor==3.6.0
python-dotenv==1.0.1
dnspython==2.7.0
